Fixup logging tests.
