+++ autopush/tests/test_main.py
@@ -1,14 +1,9 @@
-import os
 import unittest
 
-import twisted.trial
-import twisted.internet
-from mock import Mock, patch
+from mock import patch
 from moto import mock_dynamodb2
-from nose.tools import eq_
-from twisted.python import log
 
-from autopush.main import connection_main, endpoint_main, unified_setup
+from autopush.main import connection_main, endpoint_main
 
 mock_dynamodb2 = mock_dynamodb2()
 
@@ -19,27 +14,6 @@ def setUp():
 
 def tearDown():
     mock_dynamodb2.stop()
-
-
-class SentryLogTestCase(twisted.trial.unittest.TestCase):
-    def setUp(self):
-        twisted.internet.base.DelayedCall.debug = True
-        raven_patcher = patch("autopush.main.raven")
-        self.mock_raven = raven_patcher.start()
-        self.mock_client = Mock()
-        self.mock_raven.Client.return_value = self.mock_client
-
-    def tearDown(self):
-        self.mock_raven.stop()
-
-    def test_sentry_logging(self):
-        os.environ["SENTRY_DSN"] = "some_locale"
-        unified_setup()
-        eq_(len(self.mock_raven.mock_calls), 2)
-
-        log.err(Exception("eek"))
-        self.flushLoggedErrors()
-        eq_(len(self.mock_client.mock_calls), 1)
 
 
 class ConnectionMainTestCase(unittest.TestCase):
